+++ app/Models/CdrModel.php
@@ -6,4 +6,9 @@ use CodeIgniter\Model;
 class CdrModel extends Model
 {
     protected $table = 'ast_cdr';
+
+    public function search($values){
+        print_r( $values);
+        die;
+    }
 }